changes for pipeline

+++ src/candidates.py
@@ -27,7 +27,7 @@ THRESHOLD = 160
 
 
 def unet_candidates():
-    cands = glob.glob("../data/predictions_epoch16/*.png")
+    cands = glob.glob("../models/1465836586_unet/predictions_epoch36/*.png")
     #df = pd.DataFrame(columns=['seriesuid','coordX','coordY','coordZ','class'])
     data = []
     imname = ""
@@ -47,7 +47,7 @@ def unet_candidates():
         label_im, nb_labels = ndimage.label(image_eroded)
         imname3 = os.path.split(name)[1].replace('.png','')
 
-        if imname3 == "1.3.6.1.4.1.14519.5.2.1.6279.6001.112767175295249119452142211437_slice221":
+        if imname3 == "1.3.6.1.4.1.14519.5.2.1.6279.6001.124822907934319930841506266464_slice350":
             plt.imshow(label_im)
             plt.show()
 
@@ -59,31 +59,6 @@ def unet_candidates():
             blob_i = np.where(label_im==i,1,0)
             mass = center_of_mass(blob_i)
             centers.append([mass[1],mass[0]])
-
-        # label_im[label_im>0]=1
-        # contours= cv2.findContours(label_im.copy(), cv2.RETR_CCOMP, cv2.CHAIN_APPROX_SIMPLE)
-        # centers = []
-        # previous = [0,0]
-        # nrcenters = 0
-        # for cnt in contours[0]:
-        #     if len(cnt) > 1:
-        #         M = cv2.moments(cnt)
-        #         try:
-        #             cX = int(M["m10"] / M["m00"])
-        #             cY = int(M["m01"] / M["m00"])
-        #             if (cX < previous[0] - 1 or cX > previous[0] + 1) and (cY < previous[1] - 1 or cY > previous[1] + 1):
-        #                 centers.append([cX,cY])
-        #                 previous = [cX,cY]
-        #                 nrcenters+=1
-        #         except:
-        #             pass
-        #     else:
-        #         cX = cnt[0][0][0]
-        #         cY = cnt[0][0][1]
-        #         if (cX < previous[0] - 1 or cX > previous[0] + 1) and (cY < previous[1] - 1 or cY > previous[1] + 1):
-        #             centers.append([cX,cY])
-        #             previous = [cX,cY]
-        #             nrcenters+1
 
         if imname3 == "1.3.6.1.4.1.14519.5.2.1.6279.6001.112767175295249119452142211437_slice221":
             plt.subplot(211)
@@ -101,17 +76,32 @@ def unet_candidates():
                     dic = pickle.load(handle)
                     origin = dic["origin"]
                     spacing = dic["spacing"]
+            elif os.path.isfile("../data/subset8_unet/spacings/{0}.pickle".format(imname2)):
+                with open("../data/subset8_unet/spacings/{0}.pickle".format(imname2), 'rb') as handle:
+                    dic = pickle.load(handle)
+                    origin = dic["origin"]
+                    spacing = dic["spacing"]
             else:
-                _,origin,spacing=load_itk_image("../data/subset9_unet/subset9/{0}.mhd".format(imname2))
-                dic = {"origin":origin,"spacing":spacing}
-                with open('../data/subset9_unet/spacings/{0}.pickle'.format(imname2), 'wb') as handle:
-                    pickle.dump(dic, handle)
+                if os.path.isfile("../data/subset9_unet/subset9/{0}.mhd".format(imname2)):
+                    _,origin,spacing=load_itk_image("../data/subset9_unet/subset9/{0}.mhd".format(imname2))
+                    dic = {"origin":origin,"spacing":spacing}
+                    with open('../data/subset9_unet/spacings/{0}.pickle'.format(imname2), 'wb') as handle:
+                        pickle.dump(dic, handle)
+                else:
+                    _,origin,spacing=load_itk_image("../data/subset8_unet/subset8/{0}.mhd".format(imname2))
+                    dic = {"origin":origin,"spacing":spacing}
+                    with open('../data/subset8_unet/spacings/{0}.pickle'.format(imname2), 'wb') as handle:
+                        pickle.dump(dic, handle)
+
             imname = imname2
             nrimages +=1
 
         for center in centers:
-            coords = voxel_2_world([int(slice),center[1]+(512-324)*0.5,center[0]+(512-324)*0.5],origin,spacing)
-            data.append([imname2,coords[0],coords[1],coords[2],0])
+            coords = voxel_2_world([int(slice),center[1]+(512-420)*0.5,center[0]+(512-420)*0.5],origin,spacing)
+            if os.path.isfile("../data/subset9_unet/spacings/{0}.pickle".format(imname2)):
+                data.append([imname2,coords[0],coords[1],coords[2],9])
+            else:
+                data.append([imname2,coords[0],coords[1],coords[2],8])
 
         #if nrimages == 5:
         #    break
